+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Added shared README reference for each NuGet package to `Directory.Build.Props`.
 - Conslidated shared properties into `Directory.Build.Props` and removed from individual projects.
 - Removed redundant version in test project.
+- Replace xunit with xunit.v3 in test project.
+- Replace FluentAssertions with Shouldly in test project.
 - `Audacia.CodeAnalysis.Analyzers.Helpers` now referenced as private asset.
 
 ## 3.0.5 - 2025-05-01